% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_table_variable_dictionary.R
\name{create_data_table_variable_dictionary}
\alias{create_data_table_variable_dictionary}
\title{create_data_table_variable_dictionary}
\usage{
create_data_table_variable_dictionary(table_name)
}
\arguments{
\item{table_name}{string name of the table to work with}
}
\value{
a named list of standardised variable names mapped to db names
}
\description{
create_data_table_variable_dictionary
}
\examples{
create_data_table_variable_dictionary("vw_APC_SEM_001")

}
