% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospital_admissions.R
\name{create_list_of_code_column_names}
\alias{create_list_of_code_column_names}
\title{create_list_of_code_column_names}
\usage{
create_list_of_code_column_names(s, data_table_name)
}
\arguments{
\item{s}{the search strategy}

\item{data_table_name}{the name of the datatable}
}
\value{
list of variable names
}
\description{
create_list_of_code_column_names
}
\examples{
create_list_of_code_column_names("primary", "vw_APC_SEM_001")

}
