% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_medication_timeseries.R
\name{create_medication_timeseries}
\alias{create_medication_timeseries}
\title{create_medication_timeseries}
\usage{
create_medication_timeseries(
  cohort_df,
  meds_df,
  t_start,
  t_end,
  step,
  window_span,
  t_units,
  search_meds,
  merge_func = function(x) {
     mean(as.numeric(x), na.rm = T)
 }
)
}
\arguments{
\item{cohort_df}{d}

\item{meds_df}{d}

\item{t_start}{d}

\item{t_end}{d}

\item{step}{d}

\item{window_span}{d}

\item{t_units}{d}

\item{search_meds}{d}

\item{merge_func}{d}
}
\value{
medication timeseries
}
\description{
create_medication_timeseries
}
