% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survival_timings.R
\name{create_survival_timings}
\alias{create_survival_timings}
\title{create_survival_timings}
\usage{
create_survival_timings(data, outcome_groupings = NA)
}
\arguments{
\item{data}{the events data frame with columns c("nhs_number", "event_desc", "index_start", "event_start", "event_end", "last_fu")}

\item{outcome_groupings}{this must be NA, character vector, or a named list of character vectors"}
}
\value{
the timings data frame for Cox PH modelling
}
\description{
create_survival_timings
}
