% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_msoa_geo_data.R
\name{gen_msoa_geo_data}
\alias{gen_msoa_geo_data}
\title{gen_msoa_geo_data}
\usage{
gen_msoa_geo_data(msoa, values, dat = NULL)
}
\arguments{
\item{msoa}{msoa codes}

\item{values}{values to plot (numerical)}

\item{dat}{an optional data.frame with at least the col names 'msoa' and 'values'}
}
\value{
a geo dataframe ready for ploting
}
\description{
gen_msoa_geo_data
}
