% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_codes}
\alias{get_codes}
\title{get_codes}
\usage{
get_codes(file_path, append_character = NA_character_)
}
\arguments{
\item{file_path}{the file path to the csv file containing the icd-10 codes.}

\item{append_character}{a character to append to the ICD-10 codes; e.g. append_character = "-" with give "I200" and "I200-"
This is useful becuase a small subset of SUS codes have other characters appended}
}
\value{
returns a data frame with the icd-10 codes, or the ICD-10 codes plus the same codes with an appended character
}
\description{
get_codes
}
