% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hospital_admissions.R
\name{get_hospital_admissions}
\alias{get_hospital_admissions}
\title{get_hospital_admissions}
\usage{
get_hospital_admissions(
  sus_apc_obj,
  codes_list = NA,
  return_all_codes = FALSE,
  search_strategy = "primary_diagnosis",
  search_strat_adm_method = "all",
  datetime_window = lubridate::interval(NA, NA, tzone = "GMT"),
  id_list = NA,
  ccg_provider_code_list = NA,
  level = "spell",
  verbose = FALSE
)
}
\arguments{
\item{sus_apc_obj}{an R6 Database object connected to the APC_SUS database}

\item{codes_list}{list of ICD-10 diagnosis codes or OPCS procedure codes to search for}

\item{return_all_codes}{whether to return all the ICD-10 codes associated with the episode/spell, or just those in the 'codes_list'}

\item{search_strategy}{whether to look in primary diagnosis field, secondary diagnosis field, all diagnosis fields, procedure fields}

\item{search_strat_adm_method}{the admission method}

\item{datetime_window}{the beginning and end of the date window (inclusive), must be a lubridate::interval object e.g. lubridate::interval({start}, {end}, tzone = "GMT")}

\item{id_list}{vector list of character pseudo NHS ids to filter on}

\item{ccg_provider_code_list}{list of CCG provider codes, default = ALL}

\item{level}{whether to extract consultant episode or spell level data}

\item{verbose}{TRUE = print SQL query to console}
}
\value{
a data frame
}
\description{
get_hospital_admissions
}
