% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_swd_date_range.R
\name{get_swd_date_range}
\alias{get_swd_date_range}
\title{get_swd_date_range}
\usage{
get_swd_date_range(swd_att_hist, patient_ids)
}
\arguments{
\item{patient_ids}{list of patient ids}

\item{swd_att_hist_obj}{an R6 Database object connected to the SWD attributes history database}
}
\value{
a df to append to the ids
}
\description{
get_swd_date_range
}
