% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_admissions_v2.R
\name{load_admissions_v2}
\alias{load_admissions_v2}
\title{load_admissions}
\usage{
load_admissions_v2(db_conn_struct, patient_ids)
}
\arguments{
\item{db_conn_struct}{list of Database objects}

\item{patient_ids}{double, patient ids}
}
\value{
modifies the patient objects in-place
}
\description{
load_admissions
}
