% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_demographics_v2.R
\name{load_demographics_v2}
\alias{load_demographics_v2}
\title{load_demographics}
\usage{
load_demographics_v2(db_conn_struct, patient_ids)
}
\arguments{
\item{db_conn_struct}{list of Database objects}

\item{patient_ids}{list of Patient objects}
}
\value{
modifies the patient objects in-place
}
\description{
load_demographics
}
