% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_emergency_department_v2.R
\name{load_emergency_department_v2}
\alias{load_emergency_department_v2}
\title{load_emergency_department}
\usage{
load_emergency_department_v2(db_conn_struct, patient_ids)
}
\arguments{
\item{db_conn_struct}{list of Database objects}

\item{patient_ids}{double, patient ids}
}
\value{
the ED attendance tibble
}
\description{
load_emergency_department
}
