% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_swd_attributes_v2.R
\name{load_swd_attributes_v2}
\alias{load_swd_attributes_v2}
\title{load_swd_attributes_v2}
\usage{
load_swd_attributes_v2(db_conn_struct, patient_ids)
}
\arguments{
\item{db_conn_struct}{list of Database objects}

\item{patient_ids}{list of patient ids}
}
\value{
a df to append to the ids
}
\description{
load_swd_attributes_v2
}
