% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_results_v2.R
\name{load_swd_measurements_v2}
\alias{load_swd_measurements_v2}
\title{load_swd_measurements_v2}
\usage{
load_swd_measurements_v2(db_conn_struct, patient_ids)
}
\arguments{
\item{db_conn_struct}{list of Database objects}

\item{patient_ids}{list of Patient ids}
}
\value{
the swd activity data frame
}
\description{
load_swd_measurements_v2
}
