% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gp_medications.R
\name{query_gp_medications}
\alias{query_gp_medications}
\title{query_gp_medications}
\usage{
query_gp_medications(
  gp_medications_tbl,
  medications,
  processing_strategy = "any",
  time_window,
  return_vals
)
}
\arguments{
\item{gp_medications_tbl}{the table of meds}

\item{medications}{the meds to search for; NA will search for all meds}

\item{processing_strategy}{where to take each medication separate, look for any of the meds, or look for all of the meds}

\item{time_window}{in what time wiondow}

\item{return_vals}{what to return}
}
\value{
depends on the return value asked for
}
\description{
query_gp_medications
}
