% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_mortality.R
\name{query_mortality}
\alias{query_mortality}
\title{Title query_icd10_codes}
\usage{
query_mortality(
  mortality_tbl,
  time_window,
  return_vals,
  codes = NA,
  search_strat_diag = "1abc"
)
}
\arguments{
\item{mortality_tbl}{The mortality table}

\item{time_window}{The time window to search within - must be a lubridate::interval() object}

\item{return_vals}{A switch on what is being returned. If nothing found then NA is returned. Can be a vector}

\item{codes}{A vector ICD-10 codes to search for e.g. c("I210", "I219")}

\item{search_strat_diag}{The diagnosis search strategy, one out of c("primary", "secondary", "all")}
}
\value{
Depends on the return_val parameter
}
\description{
Title query_icd10_codes
}
