% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_opcs_codes.R
\name{query_opcs_codes}
\alias{query_opcs_codes}
\title{Title query_opcs_codes}
\usage{
query_opcs_codes(
  hospital_events_tbl,
  codes,
  search_strat_date,
  search_strat_adm_method = "all",
  time_window,
  return_vals
)
}
\arguments{
\item{hospital_events_tbl}{table of hospital episodes and associated data}

\item{codes}{A vector OPCS codes to search for e.g. c("K72")}

\item{search_strat_date}{The date search strategy; return the most recent date with "max", or the most historical date with "min"}

\item{search_strat_adm_method}{the admission method to filter on, on of c(...)
"exists" returns a boolean as to whether a code was found within the time window}

\item{time_window}{The time window to search within - must be a lubridate::interval() object}

\item{return_vals}{A switch on what is being returned. If nothing found then NA is returned. Can be a vector, in which case multiple colums returned
"code" returns the OPCS code
"spell_interval" returns the spell interval in which the OPCS code was found}
}
\value{
Depends on the return_val parameter
}
\description{
Title query_opcs_codes
}
