% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_swd_activity.R
\name{query_swd_activity}
\alias{query_swd_activity}
\title{query_swd_activity}
\usage{
query_swd_activity(
  swd_activity_tbl,
  provider,
  setting,
  urgency,
  speciality,
  time_window,
  return_vals
)
}
\arguments{
\item{swd_activity_tbl}{the table of swd activity info}

\item{provider}{the activity provider type to assess; one or more of c("primary_care_contact", "secondary", "community")}

\item{setting}{the activity setting to assess; one or more of c("gp", "anp", "...")}

\item{urgency}{the activity urgency flag; one or more of c(" ", " ", "...")}

\item{speciality}{the speciality to assess; one or more of c(" ", " ", "...")}

\item{time_window}{in what time wiondow}

\item{return_vals}{what to return; one or more of c(" ", " ", "...")}
}
\value{
depends on the return value asked for
}
\description{
query_swd_activity
}
