% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_swd_attributes.R
\name{query_swd_attributes}
\alias{query_swd_attributes}
\title{query_swd_attributes}
\usage{
query_swd_attributes(
  swd_attributes_tbl,
  time_window,
  search_strat_date,
  return_vals
)
}
\arguments{
\item{swd_attributes_tbl}{the table of swd attributes info}

\item{time_window}{in what time wiondow}

\item{search_strat_date}{either min or max}

\item{return_vals}{what to return; one or more of c(" ", " ", "...")}
}
\value{
depends on the return value asked for
}
\description{
query_swd_attributes
}
